Newton optimizations (presmoothing, local linear GS, overlapped communication).

+++ nts.h
@@ -42,7 +42,7 @@ struct newton
 
   int presmooth; /* presmoothing steps bound */
 
-  int refine; /* refined constraints count */
+  double refine; /* % of refined constraints */
 
   /* output */
